Nothing

+++ tdt.it.finalproject/src/main/tdt/it/finalproject/jsondata/service/ReadDollarJson.java
@@ -46,9 +46,9 @@ public class ReadDollarJson implements IReadJson {
 				JSONObject jsonObject = (JSONObject) jsonArray.get(i);
 				int id = Integer.parseInt((String) jsonObject.get("id"));
 				String name = (String) jsonObject.get("name");
-				float buyCash = (Float) jsonObject.get("buyCash");
-				float buyTransfer = (Float) jsonObject.get("buyTransfer");
-				float sellPrice = (Float) jsonObject.get("sellPrice");
+				double buyCash = (double) jsonObject.get("buyCash");
+				double buyTransfer = (double) jsonObject.get("buyTransfer");
+				double sellPrice = (double) jsonObject.get("sellPrice");
 				String date = (String) jsonObject.get("date");
 				result.add(new DollarPrice(id, name, buyCash, buyTransfer, sellPrice, date));
